% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilute.R
\name{dilute}
\alias{dilute}
\alias{dilute_to_density}
\title{Dilute a population}
\usage{
dilute(population, survival_prob)

dilute_to_density(population, d)
}
\arguments{
\item{population}{An emulsifyr population}

\item{survival_prob}{The probability of each individual surviving}
}
\value{
a population

\code{dilute_to_density} dilutes a population to a given density (approx.)
}
\description{
\code{dilute} TODO
}
\examples{
\dontrun{
p1 <- create_population(size = 1e6, ndroplets = 1e4, prop_yield = 0.5) \%>\% dilute(survival_prob = 0.01)
}
}

