% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulsify.R
\name{emulsify}
\alias{demulsify}
\alias{emulsify}
\alias{is_emulsified}
\title{Emulsify and demulsify a population}
\usage{
emulsify(population)

demulsify(population)

is_emulsified(population)
}
\arguments{
\item{population}{An emulsifyr population}
}
\value{
\code{emulsify} returns an emulsified population

\code{demulsify} TODO

\code{is_emulsified} returns a logical value indicating whether
the population is emulsified (\code{TRUE}) or not (\code{FALSE}).
}
\description{
\code{emulsify} emulsifies a population, \code{demulsify} demulsifies a population, and \code{is_emulsified} indicates whether or not a population is
emulsified
}
\examples{
\dontrun{
p1 <- create_population(size = 1e6, ndroplets = 1e4, prop_yield = 0.5)
p1_emulsified <- emulsify(population = p1)}

}
\seealso{
\code{\link{num_droplets}}
}

