% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow.R
\name{grow}
\alias{grow}
\title{Grow a population}
\usage{
grow(population)
}
\arguments{
\item{population}{}

\item{population}{An emulsifyr population}
}
\value{
a population
}
\description{
TODO
}
\examples{
\dontrun{
p1 <- create_population(size = 1e6, ndroplets = 1e4, prop_yield = 0.5)
p1_grown <- grow(population = p1)}

}

