% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{Mutate a population}
\usage{
mutate(population)
}
\arguments{
\item{population}{An emulsifyr population}
}
\value{
A thinned population
}
\description{
Mutate a population
}
\examples{
\dontrun{
p1 <- create_population(size = 1e6, ndroplets = 1e4, prop_yield = 0.5)
mutate(population = p1)
}
}

